you too

+++ code/modules/reagents/reagent_containers/spray.dm
@@ -79,7 +79,7 @@
 	D.icon += mix_color_from_reagents(D.reagents.reagent_list)
 
 	for(var/i in 1 to spray_currentrange)
-		if(!step_towards(D, A))
+		if(!step_towards(D, A) && i != 1)
 			qdel(D)
 			return
 		D.reagents.reaction(get_turf(D))
